Speedups!
Major speedups by removing call to `KMinValues.AddHash` inside the
`KMinValuesFromBytes` function.  Also, moved KMinValues.MaxSize to be
type int instead of uint64 to save on type conversions.

+++ db.go
@@ -45,7 +45,7 @@ type AddHashRequest struct {
 
 type ResizeRequest struct {
 	Key        string
-	NewSize    uint64
+	NewSize    int
 	ResultChan chan Result
 }
